requirements.txt
pycryptodome
wheel
